Refactoritzar: funcions de consulta BBDD en un sol fitxer + Schemas exportats com objecte
