+++ index.js
@@ -1,11 +1,13 @@
 const inquirer = require("inquirer");
 const { preguntasDni, preguntas } = require("./datos/preguntas");
-const { listarAnimales } = require("./listarAnimales");
-const { listarEspecieAnimales } = require("./ListarEspecieAnimales");
-const { mostrarDatosAnimal } = require("./mostrarDatosAnimal");
+const {
+  listarAnimales,
+  listarEspecieAnimales,
+  mostrarDatosAnimal,
+  adoptarAnimal,
+  cambiarNombre,
+} = require("./bd/consultas/consultasAnimales");
 const { dniEnbd } = require("./schemas/Duenyo");
-const { cambiarNombre } = require("./cambiarNombre");
-const { adoptarAnimal } = require("./adoptarAnimal");
 
 inquirer.prompt(preguntasDni).then(async (response) => {
   const { hayDuenyo, duenyo } = await dniEnbd(response.dni);
@@ -17,10 +19,10 @@ inquirer.prompt(preguntasDni).then(async (response) => {
         listarEspecieAnimales(duenyo.id, response.especie);
       } else if (response.opcion === "datosUnAnimal") {
         mostrarDatosAnimal(duenyo.id, response.chipAnimal);
-      } else if (response.opcion === "cambiarNombre") {
-        cambiarNombre(duenyo.id, response.nuevoNombre);
       } else if (response.opcion === "adoptarUnAnimal") {
         await adoptarAnimal(response.opcion, duenyo.id);
+      } else if (response.opcion === "cambiarNombre") {
+        cambiarNombre(duenyo.id, response.nuevoNombre);
       }
     });
   }
