+++ schemas/Animal.js
@@ -1,7 +1,7 @@
 const { DataTypes } = require("sequelize");
 const sequelize = require("../bd/conexion");
 const { Duenyo } = require("./Duenyo");
-const Especie = require("./Especie");
+const { Especie } = require("./Especie");
 
 const Animal = sequelize.define(
   "Animal",
@@ -45,4 +45,4 @@ Duenyo.hasMany(Animal, { foreignKey: "duenyo" });
 Animal.belongsTo(Especie, { foreignKey: "especie" });
 Especie.hasMany(Animal, { foreignKey: "especie" });
 
-module.exports = Animal;
+module.exports = { Animal };
